requirements.txt
flask>=3.0.0
line-bot-sdk>=3.0.0
apscheduler>=3.10.0
gunicorn>=21.0.0
